GQSS-962: NetworkHealthMonitor and health checking.
Currently limited only to IP resynchronization on virtual routers. This feature
allows for the removal of IPs that should not be on the virtual router, and
addition of IPs that somehow got removed. It is intended to correct IP issues
that result from race conditions when adding or removing an IP from a virtual
router.

The health checking comes in two parts: an API command checkNetworkHealth and
a monitoring thread NetworkHealthMonitor. By default the monitor runs every 3
minutes and calls the NetworkService's checkHealth method. The interval between
runs can be configured with the global setting "network.healthcheck.interval".

The API command is available to administrators only. It takes the network ID
and runs the health check as its action.

The action of health checking is somewhat synchronized. Right now it only
resyncs IP addresses. This is synchronized on a per-user basis with the
GlobalInternLock (uses UUID of account for the lock name).

The health checking is also implemented modularly following CloudStack's usual
architecture patterns. Each NetworkElement implements a resyncIPs method, but
only the VirtualRouterElement's implementation does anything currently.